fix: move profilepic file
